Restrict /db_reset for deployment

+++ backend/microservices/checkout/checkout_service/__init__.py
@@ -47,15 +47,17 @@ def create_app(ENVIRONMENT='development'):
 
     app.register_blueprint(CheckoutService.main)
 
-    # TODO: Remove method before deploying
     @app.route('/db_reset', methods=['GET'])
     def db_create():
+        if ENVIRONMENT == 'production':
+            return jsonify({'message': 'Request denied.'}), 403
+
         db.session.execute(text("DROP TABLE IF EXISTS maga_order;"))
         db.session.commit()
         with app.app_context():
             db.create_all()
 
-        return 'Tables Reset!'
+        return jsonify({'message': 'Tables reset!'}), 200
 
     return app
